DDCE-3246: Change IntegrationSpec names to avoid conflicts
