+++ it/workers/MetricsWorkerISpec.scala
@@ -17,8 +17,6 @@
 package workers
 
 import akka.stream.Materializer
-
-import java.time.{LocalDateTime, ZoneOffset}
 import com.github.tomakehurst.wiremock.client.WireMock.{any => _}
 import com.typesafe.config.ConfigFactory
 import helpers.IntegrationSpecCommonBase
@@ -38,9 +36,10 @@ import services.{ChargeReferenceService, ValidationService}
 import uk.gov.hmrc.mongo.test.DefaultPlayMongoRepositorySupport
 import utils.WireMockHelper
 
+import java.time.{LocalDateTime, ZoneOffset}
 import scala.concurrent.ExecutionContext.Implicits.global
 
-class MetricsWorkerSpec
+class MetricsWorkerISpec
     extends IntegrationSpecCommonBase
     with WireMockHelper
     with DefaultPlayMongoRepositorySupport[Declaration] {
