+++ it/workers/PaymentTimeoutWorkerISpec.scala
@@ -16,30 +16,30 @@
 
 package workers
 
+import akka.stream.Materializer
+import com.github.tomakehurst.wiremock.client.WireMock.{any => _}
 import com.typesafe.config.ConfigFactory
 import helpers.IntegrationSpecCommonBase
-
-import java.time.{LocalDateTime, ZoneOffset}
 import logger.TestLoggerAppender
 import models.ChargeReference
 import models.declarations.{Declaration, State}
+import org.mongodb.scala.model.Filters
+import org.scalatest.matchers.must.Matchers.convertToAnyMustWrapper
 import play.api.Configuration
 import play.api.inject.guice.GuiceApplicationBuilder
-import play.api.libs.json.{JsObject, Json}
-import play.api.test.Helpers.{await, running}
+import play.api.libs.json.Json
+import play.api.test.Helpers._
 import repositories.{DefaultDeclarationsRepository, DefaultLockRepository}
 import services.{ChargeReferenceService, ValidationService}
 import uk.gov.hmrc.mongo.test.DefaultPlayMongoRepositorySupport
+import utils.WireMockHelper
 
+import java.time.{LocalDateTime, ZoneOffset}
 import scala.concurrent.ExecutionContext.Implicits.global
-import play.api.test.Helpers._
-import akka.stream.Materializer
-import org.joda.time.{DateTime, DateTimeZone}
-import org.mongodb.scala.model.Filters
-import org.scalatest.matchers.must.Matchers.convertToAnyMustWrapper
 
-class DeclarationDeletionWorkerSpec
+class PaymentTimeoutWorkerISpec
     extends IntegrationSpecCommonBase
+    with WireMockHelper
     with DefaultPlayMongoRepositorySupport[Declaration] {
 
   val validationService: ValidationService           = app.injector.instanceOf[ValidationService]
@@ -58,42 +58,131 @@ class DeclarationDeletionWorkerSpec
   private lazy val builder: GuiceApplicationBuilder =
     new GuiceApplicationBuilder()
       .configure(
-        "workers.declaration-deletion-worker.interval"   -> "1 second",
-        "workers.declaration-deletion-worker.timeToHold" -> "1 minute"
+        "declarations.payment-no-response-timeout" -> "1 minute",
+        "workers.payment-timeout-worker.interval"  -> "1 second"
       )
 
-  private val journeyData: JsObject = Json.obj(
-    "euCountryCheck"        -> "greatBritain",
-    "arrivingNICheck"       -> true,
-    "isUKResident"          -> false,
-    "bringingOverAllowance" -> true
-  )
+  "A payment timeout worker" should {
 
-  private val dateTimeInPast =
-    DateTime.now().minusMinutes(5).withZone(DateTimeZone.UTC).toString("yyyy-MM-dd'T'HH:mm:ss'Z'")
-  private val dateTime       = DateTime.now().withZone(DateTimeZone.UTC).toString("yyyy-MM-dd'T'HH:mm:ss'Z'")
+    val correlationId = "fe28db96-d9db-4220-9e12-f2d267267c29"
 
-  private val dataInPast: JsObject = Json.obj(
-    "simpleDeclarationRequest" -> Json.obj(
-      "requestCommon" -> Json.obj(
-        "receiptDate" -> dateTimeInPast
-      )
-    )
-  )
+    "must log stale declarations" in {
+      await(repository.collection.deleteMany(Filters.empty()).toFuture())
 
-  private val data: JsObject = Json.obj(
-    "simpleDeclarationRequest" -> Json.obj(
-      "requestCommon" -> Json.obj(
-        "receiptDate" -> dateTime
-      )
-    )
-  )
+      TestLoggerAppender.queue.dequeueAll(_ => true)
 
-  "A declaration deletion  worker" should {
+      val app = builder.build()
 
-    val correlationId = "fk28db96-d9db-4110-9e12-f2d268541c29"
+      running(app) {
 
-    "must not log locked UnPaid records" in {
+        val declarations = List(
+          Declaration(
+            ChargeReference(0),
+            State.PendingPayment,
+            None,
+            sentToEtmp = false,
+            None,
+            correlationId,
+            None,
+            Json.obj(),
+            Json.obj(),
+            None,
+            LocalDateTime.now(ZoneOffset.UTC).minusMinutes(5)
+          ),
+          Declaration(
+            ChargeReference(1),
+            State.PaymentFailed,
+            None,
+            sentToEtmp = false,
+            None,
+            correlationId,
+            None,
+            Json.obj(),
+            Json.obj(),
+            None,
+            LocalDateTime.now(ZoneOffset.UTC).minusMinutes(5)
+          ),
+          Declaration(
+            ChargeReference(2),
+            State.PaymentCancelled,
+            None,
+            sentToEtmp = false,
+            None,
+            correlationId,
+            None,
+            Json.obj(),
+            Json.obj(),
+            None,
+            LocalDateTime.now(ZoneOffset.UTC).minusMinutes(5)
+          ),
+          Declaration(
+            ChargeReference(3),
+            State.PendingPayment,
+            None,
+            sentToEtmp = false,
+            None,
+            correlationId,
+            None,
+            Json.obj(),
+            Json.obj(),
+            None,
+            LocalDateTime.now(ZoneOffset.UTC)
+          ),
+          Declaration(
+            ChargeReference(4),
+            State.PendingPayment,
+            None,
+            sentToEtmp = false,
+            None,
+            correlationId,
+            None,
+            Json.obj(),
+            Json.obj(),
+            None,
+            LocalDateTime.now(ZoneOffset.UTC)
+          )
+        )
+
+        await(repository.collection.insertMany(declarations).toFuture())
+
+        val worker = new PaymentTimeoutWorker(
+          repository,
+          lockRepository,
+          Configuration(ConfigFactory.load(System.getProperty("config.resource")))
+        )
+
+        TestLoggerAppender.queue.dequeueAll(_ => true)
+
+        val staleDeclarations = List(
+          worker.tap.pull().futureValue.get,
+          worker.tap.pull().futureValue.get,
+          worker.tap.pull().futureValue.get
+        )
+
+        staleDeclarations
+          .map(_.chargeReference) must contain.allOf(ChargeReference(0), ChargeReference(1), ChargeReference(2))
+        staleDeclarations
+          .map(_.state)           must contain.allOf(State.PendingPayment, State.PaymentFailed, State.PaymentCancelled)
+
+        val logEvents = List(
+          TestLoggerAppender.queue.dequeue(),
+          TestLoggerAppender.queue.dequeue(),
+          TestLoggerAppender.queue.dequeue()
+        )
+
+        logEvents.map(_.getMessage) must contain.allOf(
+          "Declaration 2 is stale, deleting",
+          "Declaration 1 is stale, deleting",
+          "Declaration 0 is stale, deleting"
+        )
+
+        val remaining = repository.collection.find()
+
+        await(remaining.collect().toFuture().map(_.toList)).size mustBe 2
+      }
+    }
+
+    "must not log locked stale records" in {
       await(repository.collection.deleteMany(Filters.empty()).toFuture())
       await(lockRepository.collection.deleteMany(Filters.empty()).toFuture())
 
@@ -112,21 +201,21 @@ class DeclarationDeletionWorkerSpec
             None,
             correlationId,
             None,
-            journeyData,
-            dataInPast,
+            Json.obj(),
+            Json.obj(),
             None,
             LocalDateTime.now(ZoneOffset.UTC).minusMinutes(5)
           ),
           Declaration(
             ChargeReference(1),
-            State.Paid,
+            State.PendingPayment,
             None,
-            sentToEtmp = true,
+            sentToEtmp = false,
             None,
             correlationId,
             None,
-            journeyData,
-            dataInPast,
+            Json.obj(),
+            Json.obj(),
             None,
             LocalDateTime.now(ZoneOffset.UTC).minusMinutes(5)
           ),
@@ -138,18 +227,17 @@ class DeclarationDeletionWorkerSpec
             None,
             correlationId,
             None,
-            journeyData,
-            data,
+            Json.obj(),
+            Json.obj(),
             None,
             LocalDateTime.now(ZoneOffset.UTC)
           )
         )
 
         await(repository.collection.insertMany(declarations).toFuture())
-
         await(lockRepository.lock(0))
 
-        val worker = new DeclarationDeletionWorker(
+        val worker = new PaymentTimeoutWorker(
           repository,
           lockRepository,
           Configuration(ConfigFactory.load(System.getProperty("config.resource")))
@@ -160,7 +248,7 @@ class DeclarationDeletionWorkerSpec
       }
     }
 
-    "must lock Paid records when it processes them" in {
+    "must lock stale records when it processes them" in {
       await(repository.collection.deleteMany(Filters.empty()).toFuture())
       await(lockRepository.collection.deleteMany(Filters.empty()).toFuture())
 
@@ -171,74 +259,48 @@ class DeclarationDeletionWorkerSpec
         val declarations = List(
           Declaration(
             ChargeReference(0),
-            State.Paid,
-            None,
-            sentToEtmp = true,
-            None,
-            correlationId,
-            None,
-            journeyData,
-            dataInPast,
-            None,
-            LocalDateTime.now(ZoneOffset.UTC).minusMinutes(5)
-          ),
-          Declaration(
-            ChargeReference(1),
-            State.Paid,
+            State.PendingPayment,
             None,
             sentToEtmp = false,
             None,
             correlationId,
             None,
-            journeyData,
-            dataInPast,
+            Json.obj(),
+            Json.obj(),
+            None,
+            LocalDateTime.now(ZoneOffset.UTC).minusMinutes(5)
+          ),
+          Declaration(
+            ChargeReference(1),
+            State.PaymentFailed,
+            None,
+            sentToEtmp = false,
+            None,
+            correlationId,
+            None,
+            Json.obj(),
+            Json.obj(),
             None,
             LocalDateTime.now(ZoneOffset.UTC).minusMinutes(5)
           ),
           Declaration(
             ChargeReference(2),
-            State.Paid,
+            State.PendingPayment,
             None,
-            sentToEtmp = true,
+            sentToEtmp = false,
             None,
             correlationId,
             None,
-            journeyData,
-            data,
+            Json.obj(),
+            Json.obj(),
             None,
             LocalDateTime.now(ZoneOffset.UTC)
-          ),
-          Declaration(
-            ChargeReference(3),
-            State.Paid,
-            amendState = Some(State.Paid),
-            sentToEtmp = true,
-            amendSentToEtmp = Some(true),
-            correlationId,
-            None,
-            journeyData,
-            dataInPast,
-            None,
-            LocalDateTime.now(ZoneOffset.UTC).minusMinutes(5)
-          ),
-          Declaration(
-            ChargeReference(4),
-            State.Paid,
-            amendState = Some(State.PendingPayment),
-            sentToEtmp = true,
-            amendSentToEtmp = Some(false),
-            correlationId,
-            None,
-            journeyData,
-            dataInPast,
-            None,
-            LocalDateTime.now(ZoneOffset.UTC).minusMinutes(5)
           )
         )
 
         await(repository.collection.insertMany(declarations).toFuture())
 
-        val worker = new DeclarationDeletionWorker(
+        val worker = new PaymentTimeoutWorker(
           repository,
           lockRepository,
           Configuration(ConfigFactory.load(System.getProperty("config.resource")))
@@ -248,10 +310,8 @@ class DeclarationDeletionWorkerSpec
         worker.tap.pull().futureValue
 
         lockRepository.isLocked(0).futureValue mustEqual true
-        lockRepository.isLocked(1).futureValue mustEqual false
+        lockRepository.isLocked(1).futureValue mustEqual true
         lockRepository.isLocked(2).futureValue mustEqual false
-        lockRepository.isLocked(3).futureValue mustEqual true
-        lockRepository.isLocked(4).futureValue mustEqual false
       }
     }
 
@@ -266,53 +326,27 @@ class DeclarationDeletionWorkerSpec
         val declarations = List(
           Declaration(
             ChargeReference(0),
-            State.Paid,
+            State.PendingPayment,
             None,
-            sentToEtmp = true,
+            sentToEtmp = false,
             None,
             correlationId,
             None,
-            journeyData,
-            dataInPast,
+            Json.obj(),
+            Json.obj(),
             None,
             LocalDateTime.now(ZoneOffset.UTC).minusMinutes(5)
           ),
           Declaration(
             ChargeReference(1),
-            State.Paid,
+            State.PendingPayment,
             None,
-            sentToEtmp = true,
+            sentToEtmp = false,
             None,
             correlationId,
             None,
-            journeyData,
-            dataInPast,
-            None,
-            LocalDateTime.now(ZoneOffset.UTC).minusMinutes(5)
-          ),
-          Declaration(
-            ChargeReference(2),
-            State.Paid,
-            Some(State.Paid),
-            true,
-            Some(true),
-            correlationId,
-            None,
-            journeyData,
-            dataInPast,
-            None,
-            LocalDateTime.now(ZoneOffset.UTC).minusMinutes(5)
-          ),
-          Declaration(
-            ChargeReference(3),
-            State.Paid,
-            Some(State.Paid),
-            true,
-            Some(false),
-            correlationId,
-            None,
-            journeyData,
-            dataInPast,
+            Json.obj(),
+            Json.obj(),
             None,
             LocalDateTime.now(ZoneOffset.UTC).minusMinutes(5)
           )
@@ -320,13 +354,12 @@ class DeclarationDeletionWorkerSpec
 
         await(repository.collection.insertMany(declarations).toFuture())
 
-        val worker = new DeclarationDeletionWorker(
+        val worker = new PaymentTimeoutWorker(
           repository,
           lockRepository,
           Configuration(ConfigFactory.load(System.getProperty("config.resource")))
         )
 
-        worker.tap.pull().futureValue
         worker.tap.pull().futureValue
         worker.tap.pull().futureValue
       }
